Move tests to relevant mod
